Move method implementations
