+++ py-polars/src/interop/numpy/mod.rs
@@ -1,5 +1,3 @@
-mod to_py_df;
-mod to_py_series;
+pub mod to_numpy_df;
+pub mod to_numpy_series;
 mod utils;
-
-pub(crate) use to_py_series::{series_to_numpy, try_series_to_numpy_view};
